init: MVP Fresh plugin adding Blog functionalities
This is the first commit implementing the basic foundation of a Fresh
plugin which add Blog functionalities for the Fresh framework.

All this code is lacking tests and proper documentation.

Several important functionalities are also missing such as any form of pagination.

This is vaguely inspired by the following codebases:

 - https://github.com/denoland/fresh-blog-example
 - https://github.com/deer/fresh_blog
 - https://github.com/littlesticks/fresh-blog

Motivation:

I wanted to redo a personal website as an opportunity to dabble with Deno
and Fresh. Fresh has well matured throughout the latests releases and I
really got curious with the Plugin capabilities among other things. I
also weren't sold by the "pluggable" approach of deer/fresh_blog and
wanted to experiments with my understanding of what is feasible.